+++ client/src/components/Card/Card.jsx
@@ -3,8 +3,8 @@ import React from "react";
 const CountryCard = ({ imgflag, name, continent, population }) => {
   return (
     <div>
-      <img src={imgflag} alt="Image not Found" />
-      <p>Nmae: {name}</p>
+      <img src={imgflag} alt="Not Found" />
+      <p>Name: {name}</p>
       <p>Continente: {continent}</p>
       <p>Población: {population}</p>
     </div>
